Solved part 1
