+++ day15/attempt1b.py
